+++ CHANGELOG.md
@@ -17,6 +17,7 @@ New functionality added in a backwards-compatible manner:
 Improvements:
 * [OLMIS-6357](https://openlmis.atlassian.net/browse/OLMIS-6357): Updated LocalDatabase to reopen database connection for every action.
 * [OLMIS-6222](https://openlmis.atlassian.net/browse/OLMIS-6222): Added paginationId attribute to pagination.
+* [OLMIS-6416](https://openlmis.atlassian.net/browse/OLMIS-6416): Updated OpenlmisCachedResource getByVersionIdentities to avoid sending requests when offline.
 
 Bug fixes:
 * [OLMIS-6403](https://openlmis.atlassian.net/browse/OLMIS-6403) Fixed incorrect Start Date and Expiry Date behaviour for system notifications.